slot and requirement connection established

+++ src/main/java/com/cognizant/BDS/model/BloodRequirement.java
@@ -18,7 +18,7 @@ public class BloodRequirement {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name="req_id")
-	private int id;
+	private Long id;
 	@Column(name="blood_group")
 	private String bloodGroup;
 	@Column(table = "state",name="state")
@@ -32,7 +32,7 @@ public class BloodRequirement {
 	@Column(name="user_id")
 	private int requesterId;//user
 	
-	public BloodRequirement(int id, String bloodGroup, String state, String area, int pincode, int contactNumber,
+	public BloodRequirement(Long id, String bloodGroup, String state, String area, int pincode, int contactNumber,
 			int requesterId) {
 		super();
 		this.id = id;
@@ -49,11 +49,11 @@ public class BloodRequirement {
 	}
 
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
